refactor(fab): rename files to match tag names
