+++ packages/core/src/components/fab-button/fab-button.ios.scss
@@ -1,5 +1,5 @@
-@import "./fab";
-@import "./fab.ios.vars";
+@import "./fab-button";
+@import "./fab-button.ios.vars";
 
 // iOS FAB Button
 // --------------------------------------------------
