+++ packages/core/src/components/fab-button/fab-button.md.scss
@@ -1,5 +1,5 @@
-@import "./fab";
-@import "./fab.md.vars";
+@import "./fab-button";
+@import "./fab-button.md.vars";
 
 // Material Design FAB Button
 // --------------------------------------------------
