+++ packages/core/src/components/fab-button/fab-button.scss
@@ -1,4 +1,4 @@
-@import "./fab.vars";
+@import "./fab-button.vars";
 
 // Floating Action Buttons
 // --------------------------------------------------
