+++ packages/core/src/components/fab/test/basic/index.html
@@ -21,6 +21,7 @@
 
         <pre id="log" style="right:10px; bottom:50px; text-shadow: 0 0 2px rgba(0, 0, 0, 0.24);" slot="fixed">log</pre>
         <ion-button>Test</ion-button>
+        <ion-fab-button>FAB</ion-fab-button>
 
         <ion-fab top right edge id="fab1" slot="fixed">
           <ion-fab-button onclick="clickMainFAB('fab1')" mini class="e2eFabTopRight"><ion-icon name="add"></ion-icon></ion-fab-button>
