+++ packages/core/src/index.d.ts
@@ -26,9 +26,9 @@ export { ChipButton } from './components/chip-button/chip-button';
 export { Column } from './components/col/col';
 export { Content } from './components/content/content';
 export { Datetime } from './components/datetime/datetime';
-export { FabContainer } from './components/fab-container/fab-container';
+export { Fab } from './components/fab/fab';
 export { FabList } from './components/fab-list/fab-list';
-export { FabButton } from './components/fab/fab';
+export { FabButton } from './components/fab-button/fab-button';
 export { Footer } from './components/footer/footer';
 export { Gesture, GestureCallback, GestureDetail } from './components/gesture/gesture';
 export { PanRecognizer } from './components/gesture/recognizers';
